+++ src/openlmis-home-page-report/home-page-report.module.js
@@ -18,20 +18,13 @@
     'use strict';
 
     /**
-     * @module openlmis-home
+     * @module openlmis-home-page-report
      *
      * @description
-     * Responsible for primary navigation elements within the OpenLMIS-UI.
+     * Responsible for displaying dashboard home page report
      */
-    angular.module('openlmis-home', [
-        'openlmis-i18n',
-        'openlmis-templates',
-        'openlmis-pagination',
-        'referencedata-system-notification',
-        'openlmis-message',
-        'openlmis-home-alerts-panel',
-        'ui.router',
-        'openlmis-home-page-report'
+    angular.module('openlmis-home-page-report', [
+        'report',
+        'openlmis-superset'
     ]);
-
 })();